Man in the Middle Attack (for HTTP sites)
A python code has been written that performs a man-in-the-middle attack. Here we can see the packages from all sites with the HTTP protocol. "Arp_Poison.py" and "Packet_Listener.py" must be run separately with their arguments given.
If we want to listen to HTTPS sites; SSLStrip and dns2proxy packages must be downloaded. Then, after the installation is done in the terminal and the MitM project is started, the "sslstrip" and "python dns2proxy.py" commands should be run in separate terminals.